% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-nycflights13.R
\name{nycflights13}
\alias{copy_nycflights13}
\alias{nycflights13}
\title{Database versions of the nycflights13 data}
\usage{
copy_nycflights13(src, replace = FALSE, ...)
}
\description{
These functions cache the data from the \code{nycflights13} database in
a local database, for use in examples and vignettes. Indexes are created
to making joining tables on natural keys efficient.
}
\keyword{internal}

