% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src-netezza.R
\name{src_netezza}
\alias{src_netezza}
\alias{tbl.src_netezza}
\title{Establishes a connection to Netezza and returns a dplyr 'src' object, src_netezza, using RODBC.}
\usage{
src_netezza(dsn)

\method{tbl}{src_netezza}(src, from, ...)
}
\arguments{
\item{dsn}{The name of the ODBC DSN to use for connecting to Netezza. Must be a string.}
}
\value{
A dplyr::src object, src_netezza, to be used with dplyr functions.
}
\description{
Establishes a connection to Netezza and returns a dplyr 'src' object, src_netezza, using RODBC.
}
\examples{
\dontrun{
netezza_connection <- src_netezza(dsn="NetezzaDSN")
}
}

